refactor: extract TranscriptionMain sub-components
Split 1018-line TranscriptionMain into 5 focused sub-components:
- TranscriptionUploadTab: drag-and-drop, file preview, upload button
- TranscriptionSelectTab: uploaded files list, selection, delete
- TranscriptionControls: settings panel + start/stop buttons
- TranscriptionResultPanel: transcript/subtitles tabs, edit, clear
- TranscriptionNextActions: translate and export action cards

TranscriptionMain is now a ~340-line orchestrator composing these
sub-components via props.
